Signing info slashing testcases
